push to branch and remove hosting for now as its not structured enough to useful

+++ Kahoot.NET/StateObject.cs
@@ -15,12 +15,12 @@ public class StateObject
     /// <summary>
     /// The current message id count
     /// </summary>
-    public long id;
+    public ulong id;
 
     /// <summary>
     /// The number of acknowledgements
     /// </summary>
-    public long ack;
+    public ulong ack;
 
     /// <summary>
     /// The lag of the connection
@@ -57,6 +57,6 @@ public class StateObject
     public ExtWithTimesync<long> ExtWithTimesync => new()
     {
         Time = new() { L = l, O = o },
-        Acknowledged = ack
+        Acknowledged = (long)ack
     };
 }
